\name{WeightedPortTest-package}
\alias{WeightedPortTest-package}
\alias{WeightedPortTest}
\docType{package}
\title{Weighted Portmanteau Test procedures for Time Series Goodness-of-fit
}
\description{Two functions that implement the Weighted Portmanteau Statistics from Fisher and Gallagher (2012). The first is essentially a weighted Ljung-Box type test that can be used for fitted ARMA processes or detecting non-linear effects. The second function can be utilized to check the adequacy of a fitted ARCH process. Both are written for backward compatibility.
}
\details{
\tabular{ll}{
Package: \tab WeightedPortTest\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-03-29\cr
License: \tab GPL (>=3)\cr
LazyLoad: \tab yes\cr
}
The two functions, \code{Weighted.Box.test()} and \code{Weighted.LM.test()}, can be used in a similiar to the \code{Box.test()} function.
}
\author{
Thomas J. Fisher and Colin M. Gallagher

Maintainer: Thomas J. Fisher <fishertho@umkc.edu>
}


